+++ src/webparts/workMinders/components/tagChoice/TagChoice.module.scss
@@ -1,4 +1,4 @@
-.wm_listChoice {
+.wm_tagChoice {
   display: flex;
   flex-direction: column;
   justify-content: space-between;
