+++ src/webparts/workMinders/components/tagChoice/TagChoiceItem.module.scss
@@ -1,4 +1,4 @@
-.wm_listChoiceItem {
+.wm_tagChoiceItem {
   display: flex;
   flex-direction: row;
   justify-content: space-between;
@@ -14,8 +14,8 @@
   min-height: 16px;
 }
 
-.wm_listChoiceItemActive {
-  @extend .wm_listChoiceItem;
+.wm_tagChoiceItemActive {
+  @extend .wm_tagChoiceItem;
   color: #0078D4FF;
   box-shadow: 0 0 10px rgba(0, 0, 0, 0.1);
   font-weight: 600;
@@ -23,7 +23,7 @@
   background-image: linear-gradient(to bottom right, #0078D433, #0078D44D);
 }
 
-.wm_listChoiceItemLabel {
+.wm_tagChoiceItemLabel {
   width: 100%;
   margin: 0;
   text-align: left;
@@ -33,7 +33,7 @@
   min-height: 16px;
 }
 
-.wm_listChoiceItemHoverSection {
+.wm_tagChoiceItemHoverSection {
   display: flex;
   flex-direction: row;
   justify-content: flex-end;
@@ -44,7 +44,7 @@
   background-color: transparent;
 }
 
-.wm_listChoiceItemButton {
+.wm_tagChoiceItemButton {
   padding: 0;
   margin: 0;
   cursor: pointer;
@@ -55,14 +55,14 @@
   height: 16px;
 }
 
-.wm_listChoiceItemButton:hover {
+.wm_tagChoiceItemButton:hover {
   color: #0078d4;
 }
 
-.wm_listChoiceItemDeleteButton {
-  @extend .wm_listChoiceItemButton;
+.wm_tagChoiceItemDeleteButton {
+  @extend .wm_tagChoiceItemButton;
 }
 
-.wm_listChoiceItemDeleteButton:hover {
+.wm_tagChoiceItemDeleteButton:hover {
   color: #ff0000;
 }
